Added authorisation setup using casl package

+++ src/users/users.controller.ts
@@ -6,7 +6,8 @@ import {
   Patch,
   Param,
   Delete,
-  UseGuards,
+  Request,
+  ForbiddenException,
 } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { UsersService } from './users.service';
@@ -18,11 +19,18 @@ import {
 } from './dto';
 import { ApiExceptionResponse } from '../common/dto/api-exception-response.dto';
 import { PublicRoute } from '../common/decorators/public-route.decorator';
+import { Ability } from '@casl/ability';
+import { CaslAbilityFactory } from '../casl/casl-ability.factory/casl-ability.factory';
+import { Actions } from '../casl/enums/actions.enum';
+import { User } from './entities/user.entity';
 
 @ApiTags('users')
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   @PublicRoute()
   @ApiResponse({
@@ -31,13 +39,18 @@ export class UsersController {
     type: ApiExceptionResponse,
   })
   @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
+  create(
+    @Body() createUserDto: CreateUserDto,
+    @Request() req: Request & { user },
+  ) {
+    return this.usersService.create(createUserDto, req.user);
   }
 
   @Get()
-  findAll() {
-    return this.usersService.findAll();
+  findAll(@Request() req: Request & { user: User }) {
+    const ability: Ability = this.caslAbilityFactory.createForUser(req.user);
+    if (ability.cannot(Actions.Read, 'all')) throw new ForbiddenException();
+    return this.usersService.findAll(req.user);
   }
 
   @ApiResponse({
@@ -46,8 +59,14 @@ export class UsersController {
     type: ApiExceptionResponse,
   })
   @Get(':id')
-  findOne(@Param() findOneUserDto: FindOneUserParamsDto) {
-    return this.usersService.findOne(findOneUserDto.id);
+  async findOne(
+    @Param() findOneUserDto: FindOneUserParamsDto,
+    @Request() req: Request & { user: User },
+  ) {
+    const foundUser: User = await this.usersService.findOne(findOneUserDto.id);
+    const ability: Ability = this.caslAbilityFactory.createForUser(req.user);
+    if (ability.cannot(Actions.Read, foundUser)) throw new ForbiddenException();
+    return foundUser;
   }
 
   @ApiResponse({
@@ -56,10 +75,15 @@ export class UsersController {
     type: ApiExceptionResponse,
   })
   @Patch(':id')
-  update(
+  async update(
     @Param() params: FindOneUserParamsDto,
     @Body() updateUserDto: UpdateUserDto,
+    @Request() req: Request & { user: User },
   ) {
+    const userToUpdate = await this.usersService.findOne(params.id);
+    const ability: Ability = this.caslAbilityFactory.createForUser(req.user);
+    if (ability.cannot(Actions.Update, userToUpdate))
+      throw new ForbiddenException();
     return this.usersService.update(params.id, updateUserDto);
   }
 
@@ -69,7 +93,14 @@ export class UsersController {
     type: ApiExceptionResponse,
   })
   @Delete(':id')
-  remove(@Param() params: DeleteUserParamsDto) {
+  async remove(
+    @Param() params: DeleteUserParamsDto,
+    @Request() req: Request & { user: User },
+  ) {
+    const userToDelete = await this.usersService.findOne(params.id);
+    const ability: Ability = this.caslAbilityFactory.createForUser(req.user);
+    if (ability.cannot(Actions.Update, userToDelete))
+      throw new ForbiddenException();
     return this.usersService.remove(params.id);
   }
 }
